+++ packages/miniapp-builder-shared/src/getAppConfig.js
@@ -32,7 +32,7 @@ module.exports = (rootDir, target, nativeLifeCycleMap, subAppRoot = '') => {
 
   appConfig.routes.map(route => {
     if (subAppRoot) {
-      route.source = `${subAppRoot}/${route.source}`;
+      route.source = normalizeOutputFilePath(join(subAppRoot, route.source));
       route.subAppRoot = subAppRoot;
     }
     route.name = route.source;