+++ packages/miniapp-lifecycles/src/index.d.ts
@@ -1,0 +1,5 @@
+declare module 'miniapp-lifecycles' {
+  export function registerNativeEventListeners(Klass: any, events: string[]);
+  export function addNativeEventListener(eventName: string, callback: any);
+  export function removeNativeEventListener(eventName: string, callback: any);
+}