+++ packages/miniapp-builder-shared/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [0.2.9] - 2021-07-27
+
+### Fixed
+
+- route error in windows
+
+## [0.2.8] - 2021-07-20
+
+### Added
+
+- Support get miniapp plugin config
+
 ## [0.2.7] - 2021-06-10
 
 ### Fixed